menambahkan perhitungan jarak dengan haversineDistance

+++ database/migrations/checklist_hydrant/2025_02_21_131717_create_hydrants_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('no_hydrant');
             $table->string('location');
             $table->string('type');
-            $table->bigInteger('longitude');
-            $table->bigInteger('latitude');
+            $table->double('longitude');
+            $table->double('latitude');
             $table->timestamps();
         });
     }
